refactor magnitudes to earthquakes
